refactor: restructure python-sdk into layered architecture with Policy interface
Extract runner.py (821→593 lines) into focused modules:
- sim/policy.py: Policy protocol + RandomPolicy for pluggable action selection
- sim/config.py: RunnerConfig, AgentRuntime, SimulationOutcome dataclasses
- sim/stall_detector.py: StallDetector (draw pile stall detection)
- sim/diagnostics.py: timeout diagnostic helpers
- sim/state_utils.py: mode(), action_key(), state extraction functions

Migrate loose scripts to proper packages with CLI entry points:
- run_full_game.py → cli/run_game.py (mage-knight-run-game)
- run_seed_sweep.py → cli/run_sweep.py (mage-knight-run-sweep)
- scripts/scan_artifacts_fame.py → tools/scan_fame.py (mage-knight-scan-fame)

run_simulations() now accepts an optional policy parameter, defaulting to
RandomPolicy. Any class implementing choose_action(state, player_id,
valid_actions, rng) satisfies the Policy protocol via structural typing.
