+++ packages/python-sdk/src/mage_knight_sdk/sim/random_policy.py
@@ -1,17 +1,8 @@
 from __future__ import annotations
 
-import random
 from typing import Any
 
 from .generated_action_enumerator import CandidateAction, enumerate_valid_actions_from_state
-
-
-class RandomValidActionPolicy:
-    def choose_action(self, state: dict[str, Any], player_id: str, rng: random.Random) -> CandidateAction | None:
-        candidates = enumerate_valid_actions(state, player_id)
-        if not candidates:
-            return None
-        return rng.choice(candidates)
 
 
 def enumerate_valid_actions(state: dict[str, Any], player_id: str) -> list[CandidateAction]:
