+++ packages/python-sdk/src/mage_knight_sdk/sim/runner.py
@@ -3,14 +3,15 @@ from __future__ import annotations
 import asyncio
 import json
 import random
-from dataclasses import dataclass
 from pathlib import Path
 from typing import Any
 
 from mage_knight_sdk.client import MageKnightClient
 from mage_knight_sdk.protocol_models import ErrorMessage, StateUpdateMessage
 
-from .bootstrap import BootstrapSession, create_game, join_game
+from .bootstrap import create_game, join_game
+from .config import AgentRuntime, RunnerConfig, SimulationOutcome
+from .diagnostics import build_timeout_diagnostics
 from .invariants import (
     InvariantViolation,
     StateInvariantTracker,
@@ -18,6 +19,7 @@ from .invariants import (
     is_terminal_events,
     is_terminal_state,
 )
+from .policy import Policy, RandomPolicy
 from .random_policy import enumerate_valid_actions
 from .reporting import (
     ActionTraceEntry,
@@ -33,146 +35,37 @@ from .reporting import (
     write_failure_artifact,
     write_run_summary,
 )
+from .stall_detector import StallDetector
+from .state_utils import action_key, mode
 
 
-@dataclass(frozen=True)
-class RunnerConfig:
-    bootstrap_api_base_url: str
-    ws_server_url: str
-    player_count: int = 2
-    runs: int = 1
-    max_steps: int = 250
-    base_seed: int = 1
-    action_timeout_seconds: float = 5.0
-    artifacts_dir: str = "sim_artifacts"
-    write_failure_artifacts: bool = False
-    write_full_artifact: bool = False
-    subscribe_lobby_on_connect: bool = False
-    forced_invalid_action_step: int | None = None
-    allow_undo: bool = True
-    stall_detection_no_draw_pile_change_turns: int = 20
-
-
-@dataclass
-class AgentRuntime:
-    session: BootstrapSession
-    client: MageKnightClient
-    state_tracker: StateInvariantTracker
-    latest_state: dict[str, Any] | None = None
-    last_events: list[Any] | None = None
-    message_version: int = 0
-    last_error_message: str | None = None
-    invariant_error: str | None = None
-
-
-@dataclass
-class SimulationOutcome:
-    result: RunResult
-    trace: list[ActionTraceEntry]
-    messages: list[MessageLogEntry]
-
-
-@dataclass
-class _StallDetector:
-    no_draw_pile_change_turns: int
-    last_draw_pile_by_player: dict[str, int | None]
-    stagnant_turns_by_player: dict[str, int]
-    last_action_key: str | None = None
-
-    @classmethod
-    def create(
-        cls,
-        *,
-        no_draw_pile_change_turns: int,
-    ) -> "_StallDetector":
-        return cls(
-            no_draw_pile_change_turns=max(1, no_draw_pile_change_turns),
-            last_draw_pile_by_player={},
-            stagnant_turns_by_player={},
-        )
-
-    def observe(
-        self,
-        *,
-        step: int,
-        player_id: str,
-        action_key: str,
-        action_type: str,
-        state: dict[str, Any],
-    ) -> dict[str, Any] | None:
-        """
-        Detect stall: N consecutive END_TURN actions by same player with no draw pile change.
-        We only count END_TURN (game turns), not individual actions—a single turn can have
-        15+ actions (play cards, move, combat) with an unchanged deck, which is normal.
-        """
-        self.last_action_key = action_key
-        draw_pile_count = _draw_pile_count_for_player(state, player_id)
-        previous_draw_pile = self.last_draw_pile_by_player.get(player_id)
-        if previous_draw_pile is None or draw_pile_count != previous_draw_pile:
-            self.last_draw_pile_by_player[player_id] = draw_pile_count
-            self.stagnant_turns_by_player[player_id] = 0
-            return None
-
-        if action_type != "END_TURN":
-            return None
-
-        stagnant_turns = self.stagnant_turns_by_player.get(player_id, 0) + 1
-        self.stagnant_turns_by_player[player_id] = stagnant_turns
-        if stagnant_turns < self.no_draw_pile_change_turns:
-            return None
-
-        all_player_ids = _player_ids_from_state(state)
-        if not all_player_ids:
-            return {
-                "reason": "Stalled loop detected (draw pile count unchanged)",
-                "details": {
-                    "step": step,
-                    "playerId": player_id,
-                    "drawPileCount": draw_pile_count,
-                    "stagnantTurns": stagnant_turns,
-                    "minStagnantTurns": self.no_draw_pile_change_turns,
-                    "lastActionKey": self.last_action_key,
-                },
-            }
-
-        min_required = self.no_draw_pile_change_turns
-        if not all(
-            self.stagnant_turns_by_player.get(pid, 0) >= min_required
-            for pid in all_player_ids
-        ):
-            return None
-
-        return {
-            "reason": "Stalled loop detected (both players: draw pile count unchanged)",
-            "details": {
-                "step": step,
-                "playerId": player_id,
-                "drawPileCount": draw_pile_count,
-                "stagnantTurnsByPlayer": {
-                    pid: self.stagnant_turns_by_player.get(pid, 0)
-                    for pid in all_player_ids
-                },
-                "minStagnantTurns": self.no_draw_pile_change_turns,
-                "lastActionKey": self.last_action_key,
-            },
-        }
-
-
-async def run_simulations(config: RunnerConfig) -> tuple[list[RunResult], RunSummary]:
+async def run_simulations(
+    config: RunnerConfig,
+    policy: Policy | None = None,
+) -> tuple[list[RunResult], RunSummary]:
+    if policy is None:
+        policy = RandomPolicy()
     results: list[RunResult] = []
     for run_index in range(config.runs):
         run_seed = config.base_seed + run_index
-        outcome = await _run_single_simulation(run_index=run_index, seed=run_seed, config=config)
+        outcome = await _run_single_simulation(
+            run_index=run_index, seed=run_seed, config=config, policy=policy,
+        )
         results.append(outcome.result)
     return results, summarize(results)
 
 
-async def _run_single_simulation(run_index: int, seed: int, config: RunnerConfig) -> SimulationOutcome:
+async def _run_single_simulation(
+    run_index: int,
+    seed: int,
+    config: RunnerConfig,
+    policy: Policy,
+) -> SimulationOutcome:
     rng = random.Random(seed)
 
     trace: list[ActionTraceEntry] = []
     messages: list[MessageLogEntry] = []
-    stall_detector = _StallDetector.create(
+    stall_detector = StallDetector.create(
         no_draw_pile_change_turns=config.stall_detection_no_draw_pile_change_turns,
     )
 
@@ -231,7 +124,7 @@ async def _run_single_simulation(run_index: int, seed: int, config: RunnerConfig
                         state_update_event, timeout_seconds=config.action_timeout_seconds
                     )
                 except TimeoutError:
-                    timeout_reason, timeout_debug = _build_timeout_diagnostics(
+                    timeout_reason, timeout_debug = build_timeout_diagnostics(
                         base_reason="Timed out waiting for active player update",
                         agents=agents,
                         allow_undo=config.allow_undo,
@@ -291,7 +184,7 @@ async def _run_single_simulation(run_index: int, seed: int, config: RunnerConfig
                     messages=messages,
                 )
 
-            mode = _mode(state)
+            current_mode = mode(state)
             all_candidates = enumerate_valid_actions(state, actor.session.player_id)
 
             # If UNDO is disabled, filter it out. Some intermediate snapshots can
@@ -307,8 +200,8 @@ async def _run_single_simulation(run_index: int, seed: int, config: RunnerConfig
                         state_update_event, timeout_seconds=config.action_timeout_seconds
                     )
                 except TimeoutError:
-                    timeout_reason, timeout_debug = _build_timeout_diagnostics(
-                        base_reason=f"Timed out waiting for actionable state (mode={mode})",
+                    timeout_reason, timeout_debug = build_timeout_diagnostics(
+                        base_reason=f"Timed out waiting for actionable state (mode={current_mode})",
                         agents=agents,
                         allow_undo=config.allow_undo,
                         trace=trace,
@@ -328,10 +221,28 @@ async def _run_single_simulation(run_index: int, seed: int, config: RunnerConfig
                 continue
 
             # Sort candidates by canonical action key for reproducible choice (seed-based RNG)
-            sorted_candidates = sorted(all_candidates, key=lambda c: _action_key(c.action))
-            candidate = rng.choice(sorted_candidates)
-            candidate_keys = {_action_key(entry.action) for entry in all_candidates}
-            chosen_key = _action_key(candidate.action)
+            sorted_candidates = sorted(all_candidates, key=lambda c: action_key(c.action))
+            candidate = policy.choose_action(
+                state=state,
+                player_id=actor.session.player_id,
+                valid_actions=sorted_candidates,
+                rng=rng,
+            )
+            if candidate is None:
+                return _finish_run(
+                    config=config,
+                    run_index=run_index,
+                    seed=seed,
+                    game_id=created.game_id,
+                    outcome=OUTCOME_INVARIANT_FAILURE,
+                    steps=step,
+                    reason="Policy returned None despite valid actions being available",
+                    trace=trace,
+                    messages=messages,
+                )
+
+            candidate_keys = {action_key(entry.action) for entry in all_candidates}
+            chosen_key = action_key(candidate.action)
             if chosen_key not in candidate_keys:
                 return _finish_run(
                     config=config,
@@ -355,7 +266,7 @@ async def _run_single_simulation(run_index: int, seed: int, config: RunnerConfig
                     player_id=actor.session.player_id,
                     action=action_to_send,
                     source=candidate.source,
-                    mode=mode,
+                    mode=current_mode,
                     current_player_id=str(state.get("currentPlayerId")),
                 )
             )
@@ -371,7 +282,7 @@ async def _run_single_simulation(run_index: int, seed: int, config: RunnerConfig
                     timeout_seconds=config.action_timeout_seconds,
                 )
             except TimeoutError:
-                timeout_reason, timeout_debug = _build_timeout_diagnostics(
+                timeout_reason, timeout_debug = build_timeout_diagnostics(
                     base_reason=f"Timed out waiting for update after action {action_to_send['type']}",
                     agents=agents,
                     allow_undo=config.allow_undo,
@@ -457,7 +368,7 @@ async def _run_single_simulation(run_index: int, seed: int, config: RunnerConfig
                 stall = stall_detector.observe(
                     step=step + 1,
                     player_id=actor.session.player_id,
-                    action_key=_action_key(action_to_send),
+                    action_key=action_key(action_to_send),
                     action_type=str(action_to_send.get("type", "")),
                     state=actor_state,
                 )
@@ -480,7 +391,7 @@ async def _run_single_simulation(run_index: int, seed: int, config: RunnerConfig
 
             step += 1
 
-        max_steps_reason, max_steps_debug = _build_timeout_diagnostics(
+        max_steps_reason, max_steps_debug = build_timeout_diagnostics(
             base_reason=f"Reached max steps ({config.max_steps}) without terminal state",
             agents=agents,
             allow_undo=config.allow_undo,
@@ -670,150 +581,11 @@ def _finish_run(
     return SimulationOutcome(result=run_result, trace=trace, messages=messages)
 
 
-def _mode(state: dict[str, Any]) -> str:
-    valid_actions = state.get("validActions")
-    if not isinstance(valid_actions, dict):
-        return "unknown"
-    mode = valid_actions.get("mode")
-    if isinstance(mode, str):
-        return mode
-    return "unknown"
-
-
-def _action_key(action: dict[str, Any]) -> str:
-    return json.dumps(action, sort_keys=True, separators=(",", ":"))
-
-
-def _draw_pile_count_for_player(state: dict[str, Any], player_id: str) -> int | None:
-    players = state.get("players")
-    if not isinstance(players, list):
-        return None
-    for player in players:
-        if not isinstance(player, dict):
-            continue
-        if player.get("id") != player_id:
-            continue
-        deck_count = player.get("deckCount")
-        return int(deck_count) if isinstance(deck_count, int) else None
-    return None
-
-
-def _player_ids_from_state(state: dict[str, Any]) -> list[str]:
-    """Return player ids from state, for stall detection across all players."""
-    players = state.get("players")
-    if not isinstance(players, list):
-        return []
-    ids: list[str] = []
-    for player in players:
-        if isinstance(player, dict):
-            pid = player.get("id")
-            if isinstance(pid, str):
-                ids.append(pid)
-    return ids
-
-
-def _build_timeout_diagnostics(
-    base_reason: str,
-    agents: list[AgentRuntime],
-    allow_undo: bool,
-    trace: list[ActionTraceEntry],
-) -> tuple[str, dict[str, Any]]:
-    last_trace = trace[-1] if trace else None
-    last_action_debug: dict[str, Any] | None = None
-    if last_trace is not None:
-        last_action_debug = {
-            "step": last_trace.step,
-            "playerId": last_trace.player_id,
-            "mode": last_trace.mode,
-            "actionType": str(last_trace.action.get("type")),
-            "source": last_trace.source,
-        }
-
-    player_snapshots = [_runtime_timeout_snapshot(agent, allow_undo) for agent in agents]
-    timeout_debug = {
-        "baseReason": base_reason,
-        "lastAction": last_action_debug,
-        "players": player_snapshots,
-    }
-
-    if last_action_debug is None:
-        last_action_fragment = "none"
-    else:
-        last_action_fragment = (
-            f"step={last_action_debug['step']},player={last_action_debug['playerId']},"
-            f"mode={last_action_debug['mode']},action={last_action_debug['actionType']}"
-        )
-
-    player_fragments = [_format_timeout_player_snapshot(snapshot) for snapshot in player_snapshots]
-    players_fragment = " || ".join(player_fragments) if player_fragments else "none"
-    return (
-        f"{base_reason}; last_action={last_action_fragment}; players={players_fragment}",
-        timeout_debug,
-    )
-
-
-def _runtime_timeout_snapshot(runtime: AgentRuntime, allow_undo: bool) -> dict[str, Any]:
-    state = runtime.latest_state
-    if state is None:
-        return {
-            "playerId": runtime.session.player_id,
-            "state": "missing",
-            "messageVersion": runtime.message_version,
-            "lastError": runtime.last_error_message,
-            "invariantError": runtime.invariant_error,
-        }
-
-    mode = _mode(state)
-    current_player_id = state.get("currentPlayerId")
-    valid_actions = state.get("validActions")
-    turn = valid_actions.get("turn") if isinstance(valid_actions, dict) else None
-    turn_flags = "none"
-    if isinstance(turn, dict):
-        turn_flags = (
-            f"undo={bool(turn.get('canUndo'))},end={bool(turn.get('canEndTurn'))},"
-            f"declare_rest={bool(turn.get('canDeclareRest'))},complete_rest={bool(turn.get('canCompleteRest'))},"
-            f"rest={bool(turn.get('canRest'))},announce={bool(turn.get('canAnnounceEndOfRound'))}"
-        )
-
-    all_candidates = enumerate_valid_actions(state, runtime.session.player_id)
-    non_undo_candidates = [c for c in all_candidates if c.action.get("type") != "UNDO"]
-    effective_candidates = all_candidates if allow_undo else non_undo_candidates
-    candidate_types = sorted({str(c.action.get("type", "?")) for c in effective_candidates})
-    return {
-        "playerId": runtime.session.player_id,
-        "state": "present",
-        "mode": mode,
-        "currentPlayerId": current_player_id,
-        "messageVersion": runtime.message_version,
-        "rawActionCount": len(all_candidates),
-        "effectiveActionCount": len(effective_candidates),
-        "effectiveActionTypes": candidate_types,
-        "turnFlags": turn_flags,
-        "lastError": runtime.last_error_message,
-        "invariantError": runtime.invariant_error,
-    }
-
-
-def _format_timeout_player_snapshot(snapshot: dict[str, Any]) -> str:
-    if snapshot.get("state") == "missing":
-        return (
-            f"{snapshot['playerId']}:state=missing,msg_version={snapshot['messageVersion']},"
-            f"last_error={snapshot.get('lastError')},invariant={snapshot.get('invariantError')}"
-        )
-
-    candidate_types = snapshot.get("effectiveActionTypes")
-    candidate_preview = ",".join(candidate_types[:6]) if isinstance(candidate_types, list) and candidate_types else "none"
-    return (
-        f"{snapshot['playerId']}:mode={snapshot.get('mode')},current={snapshot.get('currentPlayerId')},"
-        f"msg_version={snapshot.get('messageVersion')},raw_actions={snapshot.get('rawActionCount')},"
-        f"effective_actions={snapshot.get('effectiveActionCount')},types={candidate_preview},"
-        f"turn={snapshot.get('turnFlags')},last_error={snapshot.get('lastError')},"
-        f"invariant={snapshot.get('invariantError')}"
-    )
-
-
-def run_simulations_sync(config: RunnerConfig) -> tuple[list[RunResult], RunSummary]:
-    return asyncio.run(run_simulations(config))
+def run_simulations_sync(
+    config: RunnerConfig,
+    policy: Policy | None = None,
+) -> tuple[list[RunResult], RunSummary]:
+    return asyncio.run(run_simulations(config, policy=policy))
 
 
 def save_summary(path: str, results: list[RunResult], summary: RunSummary) -> None:
